scraper/python_requirements.txt
browsermob-proxy==0.8.0
cryptography==2.8
fake-useragent==0.1.11
numpy==1.17.4
pandas==1.3.5
psycopg2-binary==2.9.1
python-dotenv==0.19.1
selenium==4.1.0
SQLAlchemy==1.4.28
undetected-chromedriver==3.0.6
